0.8.54.2 All units recieve damage & dynamic ones dies with style
- Style == animation

+++ Hearts Of Greed/Motor2D/Barricade.h
@@ -30,7 +30,7 @@ public:
 	void Draw(float dt);
 	void DrawSelected();
 
-	int RecieveDamage(int damage);
+	int RecieveDamage(float damage);
 
 private:
 	void Die();
